# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lectern API",
    "version": "0.1.0",
    "description": "Unified RSS + read-later reader API (MiniFlux + Readeck behind one contract)."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/search": {
      "get": {
        "operationId": "search",
        "summary": "Full-text search over owned article bodies",
        "tags": [
          "documents"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "default": 20,
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full-text search over owned article bodies",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "snippet": {
                            "type": "string"
                          },
                          "rank": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "id",
                          "snippet",
                          "rank"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "results"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/documents/{id}/content": {
      "get": {
        "operationId": "getDocumentContent",
        "summary": "Get article HTML",
        "tags": [
          "documents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get article HTML",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "html": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "html"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/documents/{id}/highlights": {
      "get": {
        "operationId": "listHighlights",
        "summary": "List highlights for a document",
        "tags": [
          "highlights"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List highlights for a document",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "highlights": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "documentId": {
                            "type": "string"
                          },
                          "text": {
                            "type": "string"
                          },
                          "note": {
                            "default": null,
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "color": {
                            "default": "yellow",
                            "type": "string",
                            "enum": [
                              "yellow",
                              "red",
                              "blue",
                              "green"
                            ]
                          },
                          "startSelector": {
                            "type": "string"
                          },
                          "startOffset": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991
                          },
                          "endSelector": {
                            "type": "string"
                          },
                          "endOffset": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "documentId",
                          "text",
                          "note",
                          "color",
                          "startSelector",
                          "startOffset",
                          "endSelector",
                          "endOffset",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "highlights"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createHighlight",
        "summary": "Create a highlight",
        "tags": [
          "highlights"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Create a highlight",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "documentId": {
                      "type": "string"
                    },
                    "text": {
                      "type": "string"
                    },
                    "note": {
                      "default": null,
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "color": {
                      "default": "yellow",
                      "type": "string",
                      "enum": [
                        "yellow",
                        "red",
                        "blue",
                        "green"
                      ]
                    },
                    "startSelector": {
                      "type": "string"
                    },
                    "startOffset": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "endSelector": {
                      "type": "string"
                    },
                    "endOffset": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "documentId",
                    "text",
                    "note",
                    "color",
                    "startSelector",
                    "startOffset",
                    "endSelector",
                    "endOffset",
                    "createdAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string"
                  },
                  "color": {
                    "default": "yellow",
                    "type": "string",
                    "enum": [
                      "yellow",
                      "red",
                      "blue",
                      "green"
                    ]
                  },
                  "note": {
                    "default": null,
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "startSelector": {
                    "type": "string"
                  },
                  "startOffset": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "endSelector": {
                    "type": "string"
                  },
                  "endOffset": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "text",
                  "color",
                  "note",
                  "startSelector",
                  "startOffset",
                  "endSelector",
                  "endOffset"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/highlights/{id}": {
      "delete": {
        "operationId": "deleteHighlight",
        "summary": "Delete a highlight",
        "tags": [
          "highlights"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Delete a highlight"
          }
        }
      }
    },
...
    "/feeds/{id}": {
      "patch": {
        "operationId": "updateFeed",
        "summary": "Update a feed (rename / move folder)",
        "tags": [
          "feeds"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Update a feed (rename / move folder)",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "feedUrl": {
                      "type": "string",
                      "format": "uri"
                    },
                    "siteUrl": {
                      "default": null,
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "uri"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "folderId": {
                      "default": null,
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "folderTitle": {
                      "default": null,
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "unreadCount": {
                      "default": 0,
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "id",
                    "title",
                    "feedUrl",
                    "siteUrl",
                    "folderId",
                    "folderTitle",
                    "unreadCount"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "folderId": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "title": {
                    "type": "string"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteFeed",
        "summary": "Unsubscribe from a feed",
        "tags": [
          "feeds"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Unsubscribe from a feed"
          }
        }
      }
    },
...
    "/feeds/import": {
      "post": {
        "operationId": "importOpml",
        "summary": "Import feeds from OPML",
        "tags": [
          "feeds"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Import feeds from OPML",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "opml": {
                    "type": "string"
                  }
                },
                "required": [
                  "opml"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
    "/documents/{id}/audio": {
      "post": {
        "operationId": "synthesizeAudio",
        "summary": "Synthesize (or return cached) read-aloud audio for a document",
        "tags": [
          "tts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Synthesize (or return cached) read-aloud audio for a document"
          }
        }
      }
    }
  }
}